refactor(data): split up views according to domain
